Change the file name to "query.ts"
